@@ -16,45 +16,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "bootstrap": "^4.1.3",
     "core-js": "^2.5.7",
     "jquery": "^3.2.1",
     "ngx-virtual-scroller": "^1.0.11",
     "popper.js": "^1.14.4",
-    "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.10.2",
-    "@angular/cli": "^7.0.2",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
-    "@types/jasmine": "~2.8.9",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.1",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.5",
     "@types/node": "~10.12.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "~5.0.1",
     "copyfiles": "^2.1.0",
-    "jasmine-core": "~3.2.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.4",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^1.3.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "mkdirp": "^0.5.1",
     "protractor": "~5.4.1",
     "rimraf": "^2.6.1",
     "ts-node": "~7.0.1",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.3"
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.5"
   }
 }
