@@ -2,24 +2,24 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { MoviesComponent } from './movies.component';
 
-describe('MoviesComponent', () => {
-  let component: MoviesComponent;
-  let fixture: ComponentFixture<MoviesComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ MoviesComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(MoviesComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//describe('MoviesComponent', () => {
+//  let component: MoviesComponent;
+//  let fixture: ComponentFixture<MoviesComponent>;
+//
+//  beforeEach(async(() => {
+//    TestBed.configureTestingModule({
+//      declarations: [ MoviesComponent ]
+//    })
+//    .compileComponents();
+//  }));
+//
+//  beforeEach(() => {
+//    fixture = TestBed.createComponent(MoviesComponent);
+//    component = fixture.componentInstance;
+//    fixture.detectChanges();
+//  });
+//
+//  it('should create', () => {
+//    expect(component).toBeTruthy();
+//  });
+//});
