@@ -2,14 +2,14 @@ import { TestBed, inject } from '@angular/core/testing';
 
 import { UsersService } from './users.service';
 
-describe('UsersService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [UsersService]
-    });
-  });
-
-  it('should be created', inject([UsersService], (service: UsersService) => {
-    expect(service).toBeTruthy();
-  }));
-});
+//describe('UsersService', () => {
+//  beforeEach(() => {
+//    TestBed.configureTestingModule({
+//      providers: [UsersService]
+//    });
+//  });
+//
+//  it('should be created', inject([UsersService], (service: UsersService) => {
+//    expect(service).toBeTruthy();
+//  }));
+//});
